+++ src/main/java/ExitCommand.java
@@ -1,0 +1,6 @@
+class ExitCommand extends Command {
+    @Override
+    public void execute(TaskList taskList, Ui ui, Storage storage) {
+        Ui.sayBye();
+    }
+}
