+++ src/main/java/UnmarkCommand.java
@@ -1,0 +1,18 @@
+class UnmarkCommand extends Command {
+    private int index;
+
+    public UnmarkCommand(int index) {
+        this.index = index;
+    }
+
+    @Override
+    public void execute(TaskList taskList, Ui ui, Storage storage) {
+        try {
+            taskList.get(index).unmarkCompleted();
+            storage.saveData(taskList);
+            Ui.unmarkTask(taskList.get(index));
+        } catch (IndexOutOfBoundsException e) {
+            Ui.taskNotFound();
+        }
+    }
+}
